cast first argument of R_alloc() to size_t everywhere; don't overflow int
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/matrix@3773 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/Matrix/src/dpoMatrix.c
@@ -73,8 +73,8 @@ SEXP dpoMatrix_rcond(SEXP obj)
     const char *ul = CHAR(STRING_ELT(uplo, 0));
 
     F77_CALL(dpocon)(ul, pdim, px, pdim, &norm, &rcond,
-		     (double *) R_alloc(3 * pdim[0], sizeof(double)),
-		     (int *) R_alloc(pdim[0], sizeof(int)),
+		     (double *) R_alloc((size_t) 3 * pdim[0], sizeof(double)),
+		     (int *) R_alloc((size_t) pdim[0], sizeof(int)),
 		     &info FCONE);
 
     UNPROTECT(4);
